refactor: deletePost api 함수 postManagementAPI.ts로 이동..
