+++ src/apis/postManagementAPI.ts
@@ -308,3 +308,31 @@ export const postRejectCollaborationApplication = async (
     throw error;
   }
 };
+
+export const deletePost = async (
+  postId: number,
+  postType: 'idea-markets' | 'request-tasks' | 'collaborations',
+): Promise<void> => {
+  const url = `${BASE_URL}/${postType}/${postId}`;
+  const API_TOKEN = localStorage.getItem('accessToken');
+
+  if (!API_TOKEN) {
+    console.error('API_TOKEN이 없습니다! 다시 로그인하세요.');
+    window.location.href = '/login/personal';
+    throw new Error('인증되지 않은 사용자입니다.');
+  }
+
+  try {
+    const response = await axios.delete(url, {
+      headers: {
+        Authorization: `Bearer ${API_TOKEN}`,
+      },
+    });
+
+    console.log(`${postType}의 ${postId}번 게시글 삭제 성공:`, response.data);
+    return response.data;
+  } catch (error) {
+    console.error('아이디어 마켓 게시글 삭제 실패:', error);
+    throw error;
+  }
+};
